build: update dependencies and fix gemspec files
This commit updates the dependencies of the project and fixes the gemspec files of the authcat-account, authcat-identity, authcat-idp, authcat-password, authcat-session, and authcat-webauthn gems to include the authcat gem as a dependency. It also updates the zeitwerk dependency to version 2.6.8.

In addition, the commit removes the bundle volume from the docker-compose file, comments out a command in the post_create.sh script, and adds the authcat-account gem to the .vscode/authcat.code-workspace file.

Finally, the commit updates the Ruby version in the authcat-account GitHub Actions workflow file to 2.6.8 and adds the rubyTestExplorer.testFramework setting to the .vscode/authcat.code-workspace file.

chore(authcat.gemspec): comment out dependencies to prevent installation issues
feat(playground): add VSCode settings to disable rubyTestExplorer test framework